Generate synthetic vectors from their sampling distributions

One way of creating synthetic data is replicating sampling distribution of the variable. This function can be applied to dataframes, numeric or character/factor vectors and obtain an object of equal length with synthetic values.

Uso

generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para default
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para numeric
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para data.frame
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

Argumentos

obj

A dataframe, numeric vector or character/factor vector.

seed

Specify seed when replication is desired.

n_news

Length of the new vectors. By default, same length as input.

Valor

An object of equal dimensions (if n_news is not specified) with synthetic values.

Ejemplos

generate_synthetic_object(mtcars)
#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  16.4     7   115    83  4.36  3.83  17.8     0     0     3     1
#>  2  17.4     6    94   128  3.08  2.94  19.0     0     0     4     2
#>  3  28.8     7   291    88  4.38  3.6   15.8     0     1     4     3
#>  4  21.9     6   291   138  3.17  3.86  20.6     0     0     3     2
#>  5  19.1     5   195   133  2.96  3.01  17.5     0     1     4     1
#>  6  10.6     4   121   169  3.15  1.98  15.1     0     0     4     5
#>  7  12       4   433   144  3.18  3.82  18.7     0     0     3     2
#>  8  10.8     7    86    67  3.6   1.78  16.6     0     0     4     2
#>  9  17.5     5   244   249  3.77  2.87  17.6     1     1     4     1
#> 10  18       6    87   211  4.05  3.43  17.3     1     0     4     1
#> # … with 22 more rows

generate_synthetic_object(mtcars$mpg)
#>  [1] 18.8 22.5 13.2 12.9 23.3 17.9 19.5 30.9 19.9 21.4 16.8 21.9 22.6 24.0 24.2
#> [16] 13.9 16.8 23.9 21.9 27.2 17.4 30.8 22.2 14.7 20.8 20.6 16.0 12.7 16.8 16.3
#> [31] 30.6 17.3

generate_synthetic_object(as.factor(mtcars$cyl))
#>  [1] 6 4 4 8 8 8 4 6 8 4 6 6 8 8 4 4 4 6 4 6 8 4 6 8 8 8 8 8 6 4 8 8
#> Levels: 4 6 8