Generate synthetic vectors from their sampling distributions

One way of creating synthetic data is replicating sampling distribution of the variable. This function can be applied to dataframes, numeric or character/factor vectors and obtain an object of equal length with synthetic values.

Uso

generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para default
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para numeric
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para data.frame
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

Argumentos

obj

A dataframe, numeric vector or character/factor vector.

seed

Specify seed when replication is desired.

n_news

Length of the new vectors. By default, same length as input.

Valor

An object of equal dimensions (if n_news is not specified) with synthetic values.

Ejemplos

generate_synthetic_object(mtcars)
#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21.8     5  74.6   215  2.82  4.07  16.3     0     0     3     4
#>  2  23       4 263.    127  4.15  3.25  19.5     1     1     4     2
#>  3  14.7     4 233.    226  4.05  1.72  16.3     0     0     4     3
#>  4  19.8     5 143.    110  3.49  2.24  20.3     0     0     3     3
#>  5  16.6     7 452.    195  3.9   1.92  18.2     1     1     4     2
#>  6  25.4     7 189      73  4.12  2.90  19.9     0     0     4     3
#>  7  17.9     6 154.    229  3.3   2.96  17.3     1     0     4     4
#>  8  17.3     5  82.9   113  3.85  5.40  21.4     1     0     3     2
#>  9  26.7     5 350.    243  3.44  3.60  16.8     1     1     4     2
#> 10  31.8     6 288.    199  3.6   3.58  19.3     0     0     4     1
#> # … with 22 more rows

generate_synthetic_object(mtcars$mpg)
#>  [1] 22.4 24.7 22.7 16.8 12.4 17.1 19.9 14.9 24.7 18.9 11.6 19.3 15.1 28.7 26.7
#> [16] 17.6 11.5 15.2 33.5 14.6 31.5 21.4 18.0 26.5 17.9 28.3 17.4 18.8 30.2 21.6
#> [31] 16.1 15.5

generate_synthetic_object(as.factor(mtcars$cyl))
#>  [1] 8 4 8 4 4 8 8 4 8 4 8 4 4 4 6 8 4 6 8 4 8 4 6 8 8 6 8 4 8 4 8 6
#> Levels: 4 6 8