Generate synthetic vectors from their sampling distributions

One way of creating synthetic data is replicating sampling distribution of the variable. This function can be applied to dataframes, numeric or character/factor vectors and obtain an object of equal length with synthetic values.

Uso

generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para default
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para numeric
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para data.frame
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

Argumentos

obj

A dataframe, numeric vector or character/factor vector.

seed

Specify seed when replication is desired.

n_news

Length of the new vectors. By default, same length as input.

Valor

An object of equal dimensions (if n_news is not specified) with synthetic values.

Ejemplos

generate_synthetic_object(mtcars)
#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  24.6     6  262.   157  4.31  3.46  17.9     1     0     4     4
#>  2  14.9     5  364.   128  3.63  1.74  17.8     1     0     4     4
#>  3  19       6  372.   155  3.22  3.84  15.8     0     0     5     2
#>  4  17.9     4  295    106  3.91  2.99  18.3     0     1     5     4
#>  5  10.4     7  133.   159  3.11  3.62  15.4     1     0     3     3
#>  6  18.3     5  155.   126  3.96  1.63  18.9     1     0     4     2
#>  7  26.7     8  362.    66  2.93  4.71  15.0     1     0     3     3
#>  8  25.9     7  109.   184  3.32  3.35  20.7     0     0     4     4
#>  9  19.2     4  419.   226  2.89  3.34  17.5     1     1     4     4
#> 10  20.9     4  368.    92  3.44  2.8   18.2     0     0     4     2
#> # … with 22 more rows

generate_synthetic_object(mtcars$mpg)
#>  [1] 16.4 25.8 21.4 12.7 19.0 23.3 30.5 14.2 21.9 27.7 32.6 15.0 17.4 17.1 12.1
#> [16] 23.6 19.4 26.7 29.1 12.7 14.5 27.3 30.6 30.6 33.2 22.9 10.7 12.7 24.6 21.2
#> [31] 19.7 21.8

generate_synthetic_object(as.factor(mtcars$cyl))
#>  [1] 8 4 8 8 6 6 4 4 6 8 8 8 8 6 8 8 6 8 4 4 6 4 4 4 8 8 4 8 8 4 6 8
#> Levels: 4 6 8