Generate synthetic vectors from their sampling distributions

One way of creating synthetic data is replicating sampling distribution of the variable. This function can be applied to dataframes, numeric or character/factor vectors and obtain an object of equal length with synthetic values.

Uso

generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para default
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para numeric
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

# S3 método para data.frame
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

Argumentos

obj

A dataframe, numeric vector or character/factor vector.

seed

Specify seed when replication is desired.

n_news

Length of the new vectors. By default, same length as input.

Valor

An object of equal dimensions (if n_news is not specified) with synthetic values.

Ejemplos

generate_synthetic_object(mtcars)
#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  15.5     7 228     208  4.07  3.09  18.6     0     0     5     2
#>  2  21.2     5 224.    151  3.78  2.26  21.9     0     1     3     4
#>  3  16.2     5 402.    152  3.24  3.39  17.2     0     1     4     2
#>  4  20.8     4 263      59  3.9   2.62  19.2     0     0     4     3
#>  5  15.2     6 388.    270  4.03  2.68  16.5     0     0     3     4
#>  6  18.8     7  84.9   112  2.85  3.36  19.1     1     0     4     1
#>  7  17.9     4 201.    264  3.47  2.15  18.4     0     0     4     1
#>  8  21.9     5 214.    215  3.91  2.22  18.9     1     0     4     5
#>  9  14.3     7 159.     82  3.94  3.82  18.2     0     0     5     1
#> 10  18.3     7 106.    196  4.37  4.35  15.8     1     0     3     2
#> # … with 22 more rows

generate_synthetic_object(mtcars$mpg)
#>  [1] 15.5 22.8 26.5 13.8 19.7 19.5 10.4 18.3 17.0 26.6 18.5 14.6 24.0 18.0 24.4
#> [16] 17.6 16.9 29.0 15.6 29.4 22.3 15.9 21.8 15.2 31.2 30.7 22.9 20.9 14.4 24.6
#> [31] 15.3 27.2

generate_synthetic_object(as.factor(mtcars$cyl))
#>  [1] 4 8 8 4 6 4 8 8 6 8 4 4 4 6 8 8 4 8 4 4 6 8 6 8 8 4 6 8 6 4 4 8
#> Levels: 4 6 8